Project Initialization
This is the initialization of the project, I've dragged the unity project into my local repository's folder location and I'm pushing it to the server.

There has been a bug with the pull requests with Unity, when you open this project you should see a save as in the asset bar in the Unity, if you don't see a triangle sprite in the center of the screen, click the save as file and it should bring you to the proper game for configuration.